+++ core/main.ts
@@ -10,7 +10,6 @@ export * from "./src/ecs/components/Actions"
 export * from "./src/ecs/entities/objects/Hitbox"
 export * from "./src/ecs/systems/core/ActionSystem"
 export * from "./src/ecs/systems/core/CommandSystem"
-export * from "./src/ecs/systems/core/ControlSystem"
 export * from "./src/ecs/systems/core/CooldownSystem"
 export * from "./src/ecs/systems/core/DamageSystem"
 export * from "./src/ecs/systems/core/ExpiresSystem"
