--- core/src/ecs/systems/core/ControlSystem.ts
@@ -1,19 +1,0 @@
-import { Controlling, Entity, SystemBuilder } from "@piggo-gg/core"
-
-export const ControlSystem: SystemBuilder = ({
-  id: "ControlSystem",
-  init: (world) => ({
-    id: "ControlSystem",
-    query: ["controlling"],
-    priority: 2,
-    onTick: (entities: Entity<Controlling>[]) => {
-      entities.forEach((entity) => {
-        const character = entity.components.controlling.getCharacter(world)
-
-        if (!character) {
-          entity.components.controlling.data.entityId = ""
-        }
-      })
-    }
-  })
-})
